Move context tests to a context dir
