+++ test/unit/context/CacheContext/parseQuery.ts
@@ -1,6 +1,6 @@
 import gql from 'graphql-tag';
 
-import { CacheContext } from '../../../src/context/CacheContext';
+import { CacheContext } from '../../../../src/context/CacheContext';
 
 describe(`context.CacheContext`, () => {
   describe(`parseQuery`, () => {
